# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Suv yetkazib berish API",
        "description": "Mobil ilovalar uchun suv yetkazib berish tizimining rasmiy API hujjatlari. Barcha endpointlar Uzbek tilida izohlangan bo‘lib, mijoz (user) va xodim (director/courier) ilovalari uchun mo‘ljallangan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Asosiy API server (Laravel backend ichidagi /api prefiksi bilan)"
        }
    ],
    "paths": {
...
        "/v1/user/orders": {
            "get": {
                "tags": [
                    "Mijoz – Buyurtmalar"
                ],
                "summary": "Mijozning barcha buyurtmalari ro‘yxati",
                "operationId": "4acaf0b5ddb1d358c53bd3504787cfa7",
                "responses": {
                    "200": {
                        "description": "Buyurtmalar ro‘yxati muvaffaqiyatli qaytarildi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mijoz – Buyurtmalar"
                ],
                "summary": "Yangi buyurtma yaratish",
                "description": "Mijoz tanlangan firma va mahsulotlar bo‘yicha yangi buyurtma yaratadi. Xizmat haqi firma balansidan ushlab qolinadi.",
                "operationId": "d3bf9b1f5e7413a90135277ad77edf21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id",
                                    "address",
                                    "items"
                                ],
                                "properties": {
                                    "company_id": {
                                        "description": "Buyurtma yuborilayotgan firma ID si",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "description": "Yetkazib berish manzili",
                                        "type": "string",
                                        "example": "Toshkent shahri, Chilonzor tumani, 1-daha"
                                    },
                                    "latitude": {
                                        "description": "Manzil koordinatasi (ixtiyoriy)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 41.3111,
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "description": "Manzil koordinatasi (ixtiyoriy)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 69.2797,
                                        "nullable": true
                                    },
                                    "payment_method": {
                                        "description": "To‘lov usuli (masalan: cash)",
                                        "type": "string",
                                        "example": "cash",
                                        "nullable": true
                                    },
                                    "items": {
                                        "description": "Buyurtma tarkibidagi mahsulotlar ro‘yxati",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "description": "Mahsulot ID si",
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "quantity": {
                                                    "description": "Mahsulot soni",
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Buyurtma muvaffaqiyatli qabul qilindi"
                    },
                    "422": {
                        "description": "Valdatsiya xatosi"
                    },
                    "500": {
                        "description": "Server xatosi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/user/orders/{id}": {
            "get": {
                "tags": [
                    "Mijoz – Buyurtmalar"
                ],
                "summary": "Mijozning bitta buyurtmasi haqida maʼlumot",
                "operationId": "9988fa8cb1f4bd67712960cb7d4f019c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buyurtma ID si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyurtma topildi va maʼlumotlar qaytarildi"
                    },
                    "404": {
                        "description": "Buyurtma topilmadi yoki ushbu mijozga tegishli emas"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/user/orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Mijoz – Buyurtmalar"
                ],
                "summary": "Mijoz tomonidan buyurtmani bekor qilish",
                "description": "Faqat holati `pending` bo‘lgan buyurtmalarni bekor qilish mumkin. Firma xizmat haqi balansga qaytariladi.",
                "operationId": "55cc1ac62d4aaae3a41ce0295079820f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buyurtma ID si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Buyurtma bekor qilindi va xizmat haqi qaytarildi"
                    },
                    "404": {
                        "description": "Buyurtma topilmadi yoki bekor qilib bo‘lmaydi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer token bilan autentifikatsiya. Format: \\\"Bearer {token}\\\"",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}